+++ src/Propulsion.Cosmos/EquinoxCosmosParser.fs
@@ -2,13 +2,13 @@
 namespace Propulsion.Cosmos
 
 open Equinox.Cosmos.Store
+open Microsoft.Azure.Documents
 #else
 namespace Propulsion.CosmosStore
 
 open Equinox.CosmosStore.Core
 #endif
 
-open Microsoft.Azure.Documents
 open Propulsion.Streams
 
 /// Maps fields in an Event within an Equinox.Cosmos V1+ Event (in a Batch or Tip) to the interface defined by Propulsion.Streams
@@ -16,9 +16,7 @@ open Propulsion.Streams
 [<RequireQualifiedAccess>]
 #if COSMOSV2 
 module EquinoxCosmosParser =
-#else
-module EquinoxCosmosStoreParser =
-#endif
+
     type Document with
         member document.Cast<'T>() =
             let tmp = Document()
@@ -29,6 +27,17 @@ module EquinoxCosmosStoreParser =
     let isEquinoxBatch (d : Document) =
         d.GetPropertyValue "p" <> null && d.GetPropertyValue "i" <> null
         && d.GetPropertyValue "n" <> null && d.GetPropertyValue "e" <> null
+#else
+module EquinoxNewtonsoftParser =
+
+    type Newtonsoft.Json.Linq.JObject with
+        member document.Cast<'T>() =
+            document.ToObject<'T>()
+
+    /// Sanity check to determine whether the Document represents an `Equinox.Cosmos` >= 1.0 based batch
+    let isEquinoxBatch (d : Newtonsoft.Json.Linq.JObject) =
+        d.ContainsKey "p" && d.ContainsKey "i" && d.ContainsKey "n" && d.ContainsKey "e"
+#endif
 
     /// Enumerates the events represented within a batch
     let enumEquinoxCosmosEvents (batch : Batch) : StreamEvent<byte[]> seq =
@@ -36,7 +45,7 @@ module EquinoxCosmosStoreParser =
         batch.e |> Seq.mapi (fun offset x -> { stream = streamName; event = FsCodec.Core.TimelineEvent.Create(batch.i+int64 offset, x.c, x.d, x.m, timestamp=x.t) })
 
     /// Collects all events with a Document [typically obtained via the CosmosDb ChangeFeed] that potentially represents an Equinox.Cosmos event-batch
-    let enumStreamEvents (d : Document) : StreamEvent<byte[]> seq =
+    let enumStreamEvents d : StreamEvent<byte[]> seq =
         if isEquinoxBatch d then d.Cast<Batch>() |> enumEquinoxCosmosEvents
         else Seq.empty